refactor: 필요 없는 핸들러매핑 삭제와 디스패처 서블릿 이동
