+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/tobe/handleradapter/ControllerHandlerAdapter.java
@@ -10,8 +10,6 @@ public class ControllerHandlerAdapter implements HandlerAdapter {
 
     @Override
     public boolean isSupport(final Object handler) {
-        System.out.println("b");
-        System.out.println(handler.getClass());
         return handler instanceof Controller;
     }
 
@@ -19,7 +17,6 @@ public class ControllerHandlerAdapter implements HandlerAdapter {
     public ModelAndView handle(final HttpServletRequest request,
                                final HttpServletResponse response,
                                final Object handler) throws Exception {
-        System.out.println("=====b======");
         final Controller controller = (Controller) handler;
         final String viewPath = controller.execute(request, response);
         return new ModelAndView(new JspView(viewPath));
